requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
black==19.10b0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.12
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
tekore==2.1.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
